% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/activityReporteR-package.R
\docType{package}
\name{activityReporteR-package}
\alias{activityReporteR}
\alias{activityReporteR-package}
\title{activityReporteR: Generate Activity Reports from REDCap and GoogleSheets Data}
\description{
Automate activity reporting that otherwise requires detailed recordkeeping and formatting. Functions are provided to download data from REDCap and GoogleSheets (since these are common sources of activity data already), and then format activities into appropriate text for reports.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/ha0ye/activityReporteR}
  \item \url{https://ha0ye.github.io/activityReporteR/}
  \item Report bugs at \url{https://github.com/ha0ye/activityReporteR/issues}
}

}
\author{
\strong{Maintainer}: Hao Ye \email{hao.ye@weecology.org} (\href{https://orcid.org/0000-0002-8630-1458}{ORCID})

}
\keyword{internal}
