% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redcap_data.R
\name{download_redcap_data}
\alias{download_redcap_data}
\title{Download a REDCap database and associated dictionary}
\usage{
download_redcap_data(
  db = c("refstats", "searches"),
  data_folder = "data",
  url = "https://redcap.ctsi.ufl.edu/redcap/api/"
)
}
\arguments{
\item{db}{name of database}

\item{data_folder}{local folder in which to store files}

\item{url}{url for the REDCap API}
}
\value{
(none)
}
\description{
Download a REDCap database and associated dictionary
}
