% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zenodo_data.R
\name{get_zenodo_token}
\alias{get_zenodo_token}
\title{Download a REDCap database and associated dictionary}
\usage{
get_zenodo_token(env_var = "zenodo_token")
}
\arguments{
\item{db}{name of database}

\item{data_folder}{local folder in which to store files}

\item{url}{url for the REDCap API}
}
\value{
(none)
}
\description{
Download a REDCap database and associated dictionary
}
