% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zenodo_data.R
\name{make_zenodo_query}
\alias{make_zenodo_query}
\title{Send a query to the Zenodo API}
\usage{
make_zenodo_query(
  path = "api/deposit/depositions",
  token = get_zenodo_token(),
  ...
)
}
\arguments{
\item{path}{the path to the endpoint}

\item{token}{access token for Zenodo API}

\item{...}{additional arguments}
}
\value{
response from the API
}
\description{
Send a query to the Zenodo API
}
